权限控制

+++ backend/controllers/ArticleController.php
@@ -9,6 +9,7 @@ use Yii;
 use common\models\Article;
 use common\models\ArticleSearch;
 use yii\web\Controller;
+use yii\web\ForbiddenHttpException;
 use yii\web\NotFoundHttpException;
 use yii\filters\VerbFilter;
 
@@ -67,6 +68,9 @@ class ArticleController extends Controller
      */
     public function actionCreate()
     {
+    	if(!Yii::$app->user->can('createPost')){
+		    throw new ForbiddenHttpException('尚无文章新增权限');
+	    }
         $model = new Article();
 
         if ($model->load(Yii::$app->request->post()) && $model->save()) {
@@ -87,6 +91,9 @@ class ArticleController extends Controller
      */
     public function actionUpdate($id)
     {
+	    if(!Yii::$app->user->can('updatePost')){
+		    throw new ForbiddenHttpException('尚无文章修改权限');
+	    }
         $model = $this->findModel($id);
 
         if ($model->load(Yii::$app->request->post()) && $model->save()) {
@@ -107,6 +114,9 @@ class ArticleController extends Controller
      */
     public function actionDelete($id)
     {
+	    if(!Yii::$app->user->can('deletePost')){
+		    throw new ForbiddenHttpException('尚无文章删除权限');
+	    }
         $this->findModel($id)->delete();
 
         return $this->redirect(['index']);
